requirements.txt
Django==5.0.6
django_phonenumber_field==7.3.0
Pillow==10.3.0
